Fix issue related to HTML escaping in the assertion
The test is here to ensure that quotes in form values are working well,
not that drivers return the raw response content without any HTML
processing before (which cannot be enforced for many of them). So the test
needs to be fine with avoiding the useless HTML escaping in case the browser
strips it.