Add PHPUnit as a dependency
Until now, our CI config relies on the simple-phpunit script of the
symfony/phpunit-bridge to install a version of PHPUnit compatible with
the PHP version. But this has one drawback: the bridge could be updated
to select new PHPUnit versions before our testsuite is compatible.
So we now use a composer dependency directly.