Add test for removal of cookie in a subpath
The previous tests only ensured that a root cookie removed in a
subfolder was applied to the child folder. It did not check that a
cookie set only in a folder is correctly deleted.

This patch adds a new test for a cookie which is set with a $path
argument other than /.