Remove usage of onNotSuccessfulTest to mark tests as skipped
Tests are not marked as skipped automatically anymore when an
UnsupportedDriverActionException is caught. Instead, drivers should
configure their unsupported tests properly in their testsuite config.
PHPUnit 10 removed support for changing the test outcome in the
onNotSuccessfulTest method of the TestCase so the previous solution will
not work anymore in the future.